Move to functional components. Separate db initialize. Add rxdb-hooks implementation.
